+++ adapter/src/main/java/me/googas/chat/adapters/bossbar/LatestBossBarAdapter.java
@@ -5,14 +5,9 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import lombok.NonNull;
-import me.googas.chat.adapters.AdaptedBossBar;
 import me.googas.chat.adapters.BossBarAdapter;
 import me.googas.chat.packet.bossbar.WrappedBarColor;
 import me.googas.chat.packet.bossbar.WrappedBarStyle;
-import org.bukkit.Bukkit;
-import org.bukkit.boss.BarColor;
-import org.bukkit.boss.BarStyle;
-import org.bukkit.boss.BossBar;
 import org.bukkit.entity.Player;
 
 public class LatestBossBarAdapter implements BossBarAdapter {
@@ -20,29 +15,34 @@ public class LatestBossBarAdapter implements BossBarAdapter {
   @NonNull private final Set<LatestAdaptedBossBar> bossBars = new HashSet<>();
 
   @Override
-  public AdaptedBossBar create(
+  public @NonNull LatestAdaptedBossBar create(
       @NonNull Player player,
       @NonNull String title,
       float progress,
       WrappedBarColor color,
       WrappedBarStyle style) {
-    BarColor baseColor = BarColor.BLUE;
-    BarStyle baseStyle = BarStyle.SOLID;
-    BossBar bossBar =
-        Bukkit.createBossBar(
-            title,
-            color == null ? baseColor : color.get().orElse(baseColor),
-            style == null ? baseStyle : style.get().orElse(baseStyle));
-    bossBar.setProgress(progress);
-    bossBar.addPlayer(player);
-    LatestAdaptedBossBar adaptedBossBar = new LatestAdaptedBossBar(bossBar, player.getUniqueId());
-    bossBars.add(adaptedBossBar);
-    return adaptedBossBar;
+    LatestAdaptedBossBar bossBar = create(player).setTitle(title).setProgress(progress);
+    if (color != null) {
+      bossBar.setColor(color);
+    }
+    if (style != null) {
+      bossBar.setStyle(style);
+    }
+    bossBars.add(bossBar);
+    return bossBar.display();
   }
 
   @Override
-  public AdaptedBossBar create(@NonNull Player player, @NonNull String title, float progress) {
-    return this.create(player, title, progress, WrappedBarColor.BLUE, WrappedBarStyle.SOLID);
+  public @NonNull LatestAdaptedBossBar create(
+      @NonNull Player player, @NonNull String title, float progress) {
+    return this.create(player, title, progress, null, null);
+  }
+
+  @Override
+  public @NonNull LatestAdaptedBossBar create(@NonNull Player player) {
+    LatestAdaptedBossBar bossBar = new LatestAdaptedBossBar(player.getUniqueId());
+    bossBars.add(bossBar);
+    return bossBar;
   }
 
   @Override