+++ adapter/src/main/java/me/googas/chat/packet/bossbar/WrappedBarColor.java
@@ -1,6 +1,5 @@
 package me.googas.chat.packet.bossbar;
 
-import lombok.Getter;
 import lombok.NonNull;
 import me.googas.reflect.Wrapper;
 import org.bukkit.boss.BarColor;
@@ -13,9 +12,14 @@ public enum WrappedBarColor implements Wrapper<BarColor> {
   YELLOW(BarColor.YELLOW),
   PURPLE(BarColor.PURPLE),
   WHITE(BarColor.WHITE);
-  @NonNull @Getter private final BarColor wrapped;
+  @NonNull private final BarColor wrapped;
 
   WrappedBarColor(@NonNull BarColor wrapped) {
     this.wrapped = wrapped;
+  }
+
+  @Override
+  public BarColor getWrapped() {
+    return this.wrapped;
   }
 }