+++ sample/src/main/java/me/googas/chat/TestCommands.java
@@ -1,0 +1,8 @@
+package me.googas.chat;
+
+import org.bukkit.entity.Player;
+
+public class TestCommands {
+
+  public void test(Player player) {}
+}